+++ Config/readconfig.go
@@ -11,7 +11,7 @@ type Config struct {
 		UserName string `yaml:"username"`
 		PassWord string `yaml:"password"`
 		HostName string `yaml:"hostname"`
-		Port     int16  `yaml:"port"`
+		Port     int    `yaml:"port"`
 		Dbname   string `yaml:"dbname"`
 	} `yaml:"database"`
 }
